--- Assets/Utilities/Editor/BuildUtils.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1746622c3d9b8424a8c2eef4ebe894f0
-timeCreated: 1613392518
