--- Assets/Utilities/Editor/EditorEvents.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 4c26a75d4a904c02a84ed36fdeeebf8e
-timeCreated: 1596718772
