+++ Assets/Utilities/Editor/Odin/Drawers/BetterEvents/BetterEventListDrawer.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using Rhinox.GUIUtils.Odin;
 using Sirenix.OdinInspector.Editor;
 using UnityEngine;
