--- Assets/Utilities/Editor/Odin/Tools/PolyCounter.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 02c552e7a3194fc685a8c1aab4ba6fff
-timeCreated: 1619536446
