--- Assets/Utilities/Editor/Odin/Windows/DependenciesManager/SubstanceToMaterial.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 81d58cc10307f9a4cb4fe0574613418c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
